+++ packages/pixer/lib/pixer.dart
@@ -1,8 +1,8 @@
 /// A fast image processing library backed by Rust
 library;
 
-export 'src/fast_image_base.dart';
-export 'src/fast_image_exception.dart';
+export 'src/pixer_base.dart';
+export 'src/pixer_exception.dart';
 export 'src/filter_type.dart';
 export 'src/image_format.dart';
 export 'src/image_metadata.dart';